pyproject.toml
[project]
name = "nonebot-plugin-anipusher"
version = "0.5.0"
description = "NoneBot plugin for pushing ani-rss/emby webhooks to QQ"
authors = [{ name = "AriadusTT", email = "980868243@qq.com" }]
license = { text = "MIT" }
readme = "README.md"
urls = { Repository = "https://github.com/AriadusTT/nonebot-plugin-AniPusher" }
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot-plugin-apscheduler",
    "nonebot_plugin_waiter",
    "aiohttp>=3.11.14",
    "DBUtils>=3.1.0",
    "fastapi>=0.115.0",
    "nonebot2>=2.4.0",
    "pydantic>=1.10.21",
    "python-dotenv>=1.0.1"
]
classifiers = [
    "Programming Language :: Python :: 3.9",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler","nonebot_plugin_waiter"]
plugin_dirs = ["nonebot-plugin-anipusher"]


[tool.ruff]
line-length = 120
select = ["F", "E", "W", "I"]  # 仅保留基础检查
ignore = ["E402"]

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]  # 替代 uv_build
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.setuptools]
packages = {find = {where = ["."]}}  # 根目录扫描
package-dir = {"" = "."}
package-data = {"*" = ["others/*.png", "others/*.jpg"]}
include-package-data = true